added demo login  functionality to demo button in the nav
